FE-Feat: escape로 모달 닫는 함수 hook으로 분리
