+++ packages/ui/src/components/common/Modal/Root.tsx
@@ -11,6 +11,7 @@ import {
   useState,
   useCallback,
 } from "react";
+import useEscapeKey from "../../../hooks/useEscapeKey";
 
 export interface ModalRootProps extends PropsWithChildren {
   open?: boolean;
@@ -28,7 +29,7 @@ const ModalContext = createContext<ModalContextProps | undefined>(undefined);
 export const useModalContext = (): ModalContextProps => {
   const context = useContext(ModalContext);
   if (!context) {
-    throw new Error("(!) Modal 컨텍스트를 호출할 수 없는 범위 입니다.");
+    throw new Error("(!) Modal 컨텍스트를 호출할 수 없는 범위입니다.");
   }
   return context;
 };
@@ -46,19 +47,11 @@ export default function ModalRoot(props: ModalRootProps): JSX.Element {
     [onOpenChange],
   );
 
-  useEffect(() => {
-    const handleKeyDown = (event: KeyboardEvent): void => {
-      if (event.key === "Escape" && open) {
-        handleOpenChange(false); // 엔터 키를 눌렀을 때 모달 닫기
-      }
-    };
-
-    window.addEventListener("keydown", handleKeyDown);
-
-    return () => {
-      window.removeEventListener("keydown", handleKeyDown);
-    };
-  }, [open, handleOpenChange]);
+  useEscapeKey(() => {
+    if (open) {
+      handleOpenChange(false);
+    }
+  }, open);
 
   useEffect(() => {
     if (open) {
